Add tests to ensure gettext_support.ex stays in sync
We don't want to create a dependency between Phoenix and installer, but
we want the Gettext helper to stay in sync since both the installers and
Phoenix generators depend on it.

In particular, both installer and Phoenix have a core_components.ex
template file that must stay in sync, and this template depends on the
`maybe_gettext/3` helper from gettext_support.ex.